Widen the debugger's addresses to 32-bit for any CPU family
CpuDebugTarget's flat memory access and its PC-aware variants took
uint16_t, and BreakpointEntry/WatchpointEntry held a uint16_t start, so a
24-bit (32016) or wider coprocessor could not be served and the memory
RPCs clamped at 0xFFFF. Widen every address and PC on the interface and
the breakpoint/watchpoint hit callbacks to uint32_t, make the entry start
fields uint32_t, and bound the memory RPCs and breakpoint/watchpoint
ranges by the descriptor's address_bits rather than a hardcoded 0xFFFF.
The 6502 implementations -- the coprocessor runner and HostDebugTarget --
truncate to their 16-bit map internally, so nothing changes for a 6502.

Extend the stub-family test with a 24-bit CPU: a read, a write and a
breakpoint above 0xFFFF are all served.

+++ src/core/include/beebium/CpuBinding.hpp
@@ -36,7 +36,7 @@ concept HasPcAwareMemory = requires(T& m, uint16_t addr, uint16_t pc, uint8_t va
 using CpuInstructionCallback = std::function<void(uint16_t pc)>;
 
 // Watchpoint hit callback type for CpuBinding inline checking
-using CpuWatchpointHitCallback = std::function<void(const WatchpointEntry& wp, uint16_t addr, uint8_t value, bool is_write)>;
+using CpuWatchpointHitCallback = std::function<void(const WatchpointEntry& wp, uint32_t addr, uint8_t value, bool is_write)>;
 
 // CpuBinding wraps M6502 + MemoryPolicy for clock subscription.
 //
